Move init helper functions to init.go file
